refactor(gateway): move gateway transformation into separate file
