+++ pkg/plugins/runtime/k8s/controllers/gatewayapi/gateway_controller.go
@@ -2,7 +2,6 @@ package gatewayapi
 
 import (
 	"context"
-	"fmt"
 
 	"github.com/go-logr/logr"
 	"github.com/pkg/errors"
@@ -10,7 +9,6 @@ import (
 	kube_apierrs "k8s.io/apimachinery/pkg/api/errors"
 	kube_meta "k8s.io/apimachinery/pkg/apis/meta/v1"
 	kube_runtime "k8s.io/apimachinery/pkg/runtime"
-	kube_schema "k8s.io/apimachinery/pkg/runtime/schema"
 	kube_ctrl "sigs.k8s.io/controller-runtime"
 	kube_client "sigs.k8s.io/controller-runtime/pkg/client"
 	kube_controllerutil "sigs.k8s.io/controller-runtime/pkg/controller/controllerutil"
@@ -21,9 +19,6 @@ import (
 	mesh_k8s "github.com/kumahq/kuma/pkg/plugins/resources/k8s/native/api/v1alpha1"
 	k8s_registry "github.com/kumahq/kuma/pkg/plugins/resources/k8s/native/pkg/registry"
 	"github.com/kumahq/kuma/pkg/plugins/runtime/k8s/containers"
-	k8s_util "github.com/kumahq/kuma/pkg/plugins/runtime/k8s/util"
-	util_k8s "github.com/kumahq/kuma/pkg/util/k8s"
-	"github.com/kumahq/kuma/pkg/plugins/runtime/k8s/controllers/gatewayapi/policy"
 )
 
 // GatewayReconciler reconciles a GatewayAPI Gateway object.
@@ -79,110 +74,6 @@ func (r *GatewayReconciler) Reconcile(ctx context.Context, req kube_ctrl.Request
 	}
 
 	return kube_ctrl.Result{}, nil
-}
-
-type ListenerConditions map[gatewayapi.SectionName][]kube_meta.Condition
-
-func (r *GatewayReconciler) gapiToKumaGateway(
-	ctx context.Context, gateway *gatewayapi.Gateway,
-) (*mesh_proto.Gateway, ListenerConditions, error) {
-	var listeners []*mesh_proto.Gateway_Listener
-
-	listenerConditions := ListenerConditions{}
-
-	for _, l := range gateway.Spec.Listeners {
-		listener := &mesh_proto.Gateway_Listener{
-			Port: uint32(l.Port),
-			Tags: map[string]string{
-				// gateway-api routes are configured using direct references to
-				// Gateways, so just create a tag specifically for this listener
-				mesh_proto.ListenerTag: string(l.Name),
-			},
-		}
-
-		if protocol, ok := mesh_proto.Gateway_Listener_Protocol_value[string(l.Protocol)]; ok {
-			listener.Protocol = mesh_proto.Gateway_Listener_Protocol(protocol)
-		} else if l.Protocol != "" {
-			// TODO admission webhook should prevent this
-			listenerConditions[l.Name] = append(listenerConditions[l.Name],
-				kube_meta.Condition{
-					Type:    string(gatewayapi.ListenerConditionReady),
-					Status:  kube_meta.ConditionFalse,
-					Reason:  string(gatewayapi.ListenerReasonInvalid),
-					Message: fmt.Sprintf("unexpected protocol %s", l.Protocol),
-				},
-			)
-			continue
-		}
-
-		listener.Hostname = "*"
-		if l.Hostname != nil {
-			listener.Hostname = string(*l.Hostname)
-		}
-
-		refsPermitted := true
-
-		for _, certRef := range l.TLS.CertificateRefs {
-			ref := policy.PolicyReferenceSecret(policy.FromGatewayIn(gateway.Namespace), *certRef)
-
-			permitted, err := policy.IsReferencePermitted(ctx, r.Client, ref)
-			if err != nil {
-				return nil, nil, err
-			}
-
-			refsPermitted = refsPermitted && permitted
-
-			var resolvedCondition []kube_meta.Condition
-
-			if permitted {
-				resolvedCondition = []kube_meta.Condition{
-					{
-						Type:   string(gatewayapi.ListenerConditionResolvedRefs),
-						Status: kube_meta.ConditionTrue,
-						Reason: string(gatewayapi.ListenerReasonResolvedRefs),
-					},
-					{
-						Type:   string(gatewayapi.ListenerConditionReady),
-						Status: kube_meta.ConditionTrue,
-						Reason: string(gatewayapi.ListenerConditionReady),
-					},
-				}
-			} else {
-				targetGK := kube_schema.GroupKind{Group: string(*certRef.Group), Kind: string(*certRef.Kind)}
-				resolvedCondition = []kube_meta.Condition{
-					{
-						Type:    string(gatewayapi.ListenerConditionResolvedRefs),
-						Status:  kube_meta.ConditionFalse,
-						Reason:  string(gatewayapi.ListenerReasonRefNotPermitted),
-						Message: fmt.Sprintf("reference to %s %s/%s not permitted by any ReferencePolicy", targetGK.String(), ref.ToNamespace(), certRef.Name),
-					},
-					{
-						Type:    string(gatewayapi.ListenerConditionReady),
-						Status:  kube_meta.ConditionFalse,
-						Reason:  string(gatewayapi.ListenerReasonInvalid),
-						Message: "unable to resolve refs",
-					},
-				}
-			}
-
-			listenerConditions[l.Name] = append(listenerConditions[l.Name], resolvedCondition...)
-		}
-
-		if refsPermitted {
-			listeners = append(listeners, listener)
-		}
-	}
-
-	match := serviceTagForGateway(kube_client.ObjectKeyFromObject(gateway))
-
-	return &mesh_proto.Gateway{
-		Selectors: []*mesh_proto.Selector{
-			{Match: match},
-		},
-		Conf: &mesh_proto.Gateway_Conf{
-			Listeners: listeners,
-		},
-	}, listenerConditions, nil
 }
 
 func (r *GatewayReconciler) createOrUpdateInstance(ctx context.Context, gateway *gatewayapi.Gateway) (*mesh_k8s.GatewayInstance, error) {
